Move shared view to directory
