+++ Multiplatform/Views/Shared/LKTextField.swift
@@ -16,46 +16,6 @@
 
 import SwiftUI
 
-// Default button style for this example
-struct LKButton: View {
-    let title: String
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action,
-               label: {
-                   Text(title.uppercased())
-                       .fontWeight(.bold)
-                       .padding(.horizontal, 12)
-                       .padding(.vertical, 10)
-               })
-               .background(Color.lkRed)
-               .cornerRadius(8)
-    }
-}
-
-#if os(iOS)
-extension LKTextField.`Type` {
-    func toiOSType() -> UIKeyboardType {
-        switch self {
-        case .URL: return .URL
-        case .ascii: return .asciiCapable
-        default: return .default
-        }
-    }
-}
-#endif
-
-#if os(macOS)
-// Avoid showing focus border around textfield for macOS
-extension NSTextField {
-    override open var focusRingType: NSFocusRingType {
-        get { .none }
-        set {}
-    }
-}
-#endif
-
 struct LKTextField: View {
     enum `Type` {
         case `default`
@@ -94,3 +54,25 @@ struct LKTextField: View {
         }.frame(maxWidth: .infinity)
     }
 }
+
+#if os(iOS)
+extension LKTextField.`Type` {
+    func toiOSType() -> UIKeyboardType {
+        switch self {
+        case .URL: return .URL
+        case .ascii: return .asciiCapable
+        default: return .default
+        }
+    }
+}
+#endif
+
+#if os(macOS)
+// Avoid showing focus border around textfield for macOS
+extension NSTextField {
+    override open var focusRingType: NSFocusRingType {
+        get { .none }
+        set {}
+    }
+}
+#endif
